[C] Move Manage view to views/entities
