+++ components/views/entities/index.ts
@@ -1,0 +1,3 @@
+export { default as Manage } from "./Manage";
+export { default as SubcollectionList } from "./SubcollectionList";
+export { default as SubitemList } from "./SubitemList";
